refactor Tx
