+++ lib/Echidna/Solidity.hs
@@ -33,10 +33,12 @@ import System.IO                  (openFile, IOMode(..))
 import System.Exit                (ExitCode(..))
 import System.Directory           (findExecutable)
 
-import Echidna.ABI         (SolSignature, stripBytecodeMetadata, fallback)
-import Echidna.Exec        (execTx)
-import Echidna.RPC         (loadEthenoBatch)
-import Echidna.Transaction (TxConf, TxCall(SolCreate), Tx(..), World(..))
+import Echidna.ABI             (stripBytecodeMetadata, fallback)
+import Echidna.Exec            (execTx)
+import Echidna.RPC             (loadEthenoBatch)
+import Echidna.Types.Signature (SolSignature)
+import Echidna.Types.Tx        (TxConf, TxCall(SolCreate), Tx(..))
+import Echidna.Types.World     (World(..))
 
 import EVM hiding (contracts)
 import qualified EVM (contracts)
