+++ src/test/Spec.hs
@@ -11,12 +11,13 @@ import EVM.ABI (AbiValue(..))
 import EVM.Types (Addr)
 import qualified EVM.Concrete(Word(..))
 
-import Echidna.ABI (SolCall, mkGenDict)
+import Echidna.ABI (mkGenDict)
 import Echidna.Types.Campaign (Campaign, CampaignConf(..), TestState(..), tests, gasInfo, testLimit, shrinkLimit, knownCoverage, corpus, coverage)
 import Echidna.Campaign (campaign)
 import Echidna.Config (EConfig, EConfigWithUsage(..), _econfig, defaultConfig, parseConfig, sConf, cConf)
 import Echidna.Solidity
-import Echidna.Transaction (TxCall(..), Tx(..), call)
+import Echidna.Types.Signature (SolCall)
+import Echidna.Types.Tx (TxCall(..), Tx(..), call)
 
 import Data.Aeson (encode, decode)
 import Control.Lens
